feat: Update UI components and configurations for improved layout and functionality

+++ src/UIScreen.cpp
@@ -10,7 +10,7 @@
  *  Automatikusan teljes képernyő méretet használ (0,0 tft.width(), tft.height()).
  * Az UIContainerComponent konstruktor hívása után a név beállítása történik.
  */
-UIScreen::UIScreen(const char *name) : UIContainerComponent({0, 0, tft.width(), tft.height()}), name(name) {}
+UIScreen::UIScreen(const char *name) : UIContainerComponent({0, 0, (uint16_t)tft.width(), (uint16_t)tft.height()}), name(name) {}
 
 // ================================
 // UIComponent Override Methods - Event Handling és Rendering
